+++ src/js/gutenberg-overrides/core-blocks/gallery/gallery-image.js
@@ -2,6 +2,7 @@
  * External Dependencies
  */
 import React from 'react';
+import { reduce } from 'lodash';
 import classnames from 'classnames';
 
 /**
@@ -15,8 +16,8 @@ import { withSelect } from '@wordpress/data';
 import { RichText } from '@wordpress/editor';
 
 class GalleryImage extends Component {
-	constructor() {
-		super( ...arguments );
+	constructor(props) {
+		super(props);
 
 		this.onImageClick = this.onImageClick.bind( this );
 		this.onSelectCaption = this.onSelectCaption.bind( this );
@@ -83,10 +84,21 @@ class GalleryImage extends Component {
 				captionSelected: false,
 			} );
 		}
+
+		if ( image && image !== prevProps.image && image.data ) {
+			const data = reduce( image.data, ( result, value, key ) => {
+				key = key.replace( '_', '-' );
+				result[ `data-${ key }` ] = value;
+
+				return result;
+			}, {} );
+
+			this.props.setAttributes( { data } );
+		}
 	}
 
 	render() {
-		const { url, alt, id, linkTo, link, isSelected, caption, onRemove, setAttributes } = this.props;
+		const { url, alt, id, linkTo, link, data, isSelected, caption, onRemove, setAttributes } = this.props;
 
 		let href;
 
@@ -102,7 +114,7 @@ class GalleryImage extends Component {
 		// Disable reason: Image itself is not meant to be
 		// interactive, but should direct image selection and unfocus caption fields
 		// eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions, jsx-a11y/click-events-have-key-events
-		const img = url ? <img src={ url } alt={ alt } data-id={ id } onClick={ this.onImageClick } /> : <Spinner />;
+		const img = url ? <img src={ url } alt={ alt } data-id={ id } onClick={ this.onImageClick } { ...data } /> : <Spinner />;
 
 		const className = classnames( {
 			'is-selected': isSelected,