+++ src/js/gutenberg-overrides/core-blocks/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `data` attribute to `cover-image` block and change *onSelectImage* function to set `data` atributes (`cover-image/index.js`)
 
-- `data` attribute to `gallery` block and **GalleryImage** (`gallery/index.js`, `gallery/edit.js` and `gallery/gallery-image.js`)
+- `data` attribute to `gallery` on `save` (`gallery/index.js`)
+
+- `data` attribute to **GalleryImage** (`gallery/edit.js`)
+
+- `data` attribute to be set on `componentDidUpdate` event and then rendered (`gallery/gallery-image.js`)
 
 - `data` attribute to `image` block and change *onSelectImage* function to set `data` atributes (`image/index.js` and `image/edit.js`)